requirements.txt
flask==3.1.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
Pillow==11.1.0
pytesseract==0.3.13
